refector:add UserResponseDto

+++ hello-nest/src/users/users.service.ts
@@ -6,6 +6,7 @@ import {
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 import { User } from './entities/user.entity';
+import { UserResponseDto } from './dto/user-response.dto';
 
 @Injectable()
 export class UsersService {
@@ -14,11 +15,12 @@ export class UsersService {
     private readonly usersRepository: Repository<User>,
   ) {}
 
-  async findAll(): Promise<User[]> {
-    return this.usersRepository.find();
+  async findAll(): Promise<UserResponseDto[]> {
+    const users: User[] = await this.usersRepository.find();
+    return users.map((user) => new UserResponseDto(user));
   }
 
-  async findById(id: number): Promise<User> {
+  async findById(id: number): Promise<UserResponseDto> {
     const user = await this.usersRepository.findOne({
       where: { id },
     });
@@ -27,19 +29,23 @@ export class UsersService {
       throw new NotFoundException('User not found');
     }
 
-    return user;
+    return new UserResponseDto(user);
   }
 
   async findByEmail(email: string): Promise<User | null> {
-    return this.usersRepository.findOne({
+    const user = await this.usersRepository.findOne({
       where: { email },
     });
+
+    return user?? null;
   }
 
-  async findByUsername(username: string): Promise<User | null> {
-    return this.usersRepository.findOne({
+  async findByUsername(username: string): Promise<UserResponseDto | null> {
+    const user = await this.usersRepository.findOne({
       where: { username },
     });
+
+    return user ? new UserResponseDto(user) : null;
   }
 
   async create(userData: Partial<User>): Promise<User> {
@@ -61,17 +67,29 @@ export class UsersService {
   }
 
   async update(id: number, userData: Partial<User>): Promise<User> {
-    const user = await this.findById(id);
+    const user = await this.usersRepository.findOne({
+      where: { id },
+    });
+
+    if (!user) {
+      throw new NotFoundException('User not found');
+    }
 
     Object.assign(user, userData);
 
     return this.usersRepository.save(user);
   }
 
-  async remove(id: number): Promise<void> {
-    const user = await this.findById(id);
+  async remove(id: number) {
+    const user = await this.usersRepository.findOne({
+      where: { id },
+    });
 
-    await this.usersRepository.remove(user);
+    if (!user) {
+      throw new NotFoundException('User not found');
+    }
+
+    return this.usersRepository.remove(user);
   }
 
   async getProfile(userId: number){
